#2420 - Disable plugin manager by default
- Add feature flag for plugin manager
