+++ inception/inception-plugin-manager/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+de.tudarmstadt.ukp.clarin.webanno.plugin.config.PluginManagerAutoConfiguration
